refactor: Rename `lib/plugins/aws/deploy/lib/checkForChanges.js`
