+++ lib/plugins/aws/deploy/index.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const writeServiceOutputs = require('../../../cli/write-service-outputs');
 const extendedValidate = require('./lib/extendedValidate');
 const setBucketName = require('../lib/set-bucket-name');
-const checkForChanges = require('./lib/checkForChanges');
+const checkForChanges = require('./lib/check-for-changes');
 const monitorStack = require('../lib/monitor-stack');
 const checkIfBucketExists = require('../lib/check-if-bucket-exists');
 const getCreateChangeSetParams = require('../lib/get-create-change-set-params');
