+++ lib/plugins/aws/deploy/lib/check-for-changes.js
